RunningDistance

Documentation

class Distance d where

A generic class that defines all the units of measure in running. The base unit of meters was chosen because Metric is orders more sane than the English system

Methods

toMeters :: d -> Meter

Converts the given unit of distance into meters

fromMeters :: Meter -> d

Converts the given number of meters into the unit of measure

toFloat :: d -> Float

Yields the given distance as a float

data Microinch

Unit Definitions

English, Ascending

data Mil

data Line

data Barleycorn

data Finger

data Inch

data Nail

data Palm

data Foot

data Cubit

data Yard

data Ell

data Fathom

data Perch

data Rod

data Chain

data Furlong

data Mile

data Meter

data Kilometer

addDist :: (Distance a, Distance b, Distance c) => a -> b -> c

Functions

Add two distances together to get a common distance

subDist :: (Distance a, Distance b, Distance c) => a -> b -> c

Find the difference between two distcances

convertDist :: (Distance a, Distance b) => a -> b

Convert one distance to another.